requirements.txt
absl-py==1.1.0
action-msgs==1.2.1
action-tutorials-interfaces==0.20.5
action-tutorials-py==0.20.5
actionlib-msgs==4.2.4
ale-py==0.7.5
ament-cmake-test==1.3.11
ament-copyright==0.12.11
ament-cppcheck==0.12.11
ament-cpplint==0.12.11
ament-flake8==0.12.11
ament-index-python==1.4.0
ament-lint==0.12.11
ament-lint-cmake==0.12.11
ament-package==0.14.0
ament-pep257==0.12.11
ament-uncrustify==0.12.11
ament-xmllint==0.12.11
angles==1.15.0
astroid==2.12.10
asttokens==3.0.0
astunparse==1.6.3
attrs==22.1.0
AutoROM==0.4.2
AutoROM.accept-rom-license==0.4.2
av==10.0.0
blis==0.7.7
builtin-interfaces==1.2.1
cachetools==5.2.0
catalogue==2.0.7
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cloudpickle==2.1.0
colorlog==6.7.0
composition-interfaces==1.2.1
coverage==6.4.4
cupy-cuda113==10.5.0
cv-bridge==3.2.1
cycler==0.11.0
cymem==2.0.6
decorator==5.2.1
demo-nodes-py==0.20.5
diagnostic-msgs==4.2.4
dill==0.3.5.1
dm-env==1.6
dm-tree==0.1.9
dm_control==1.0.28
docker-pycreds==0.4.0
docstring_parser==0.16
domain-coordinator==0.10.0
-e git+https://github.com/ROBOTIS-GIT/DynamixelSDK.git@3450c7078917b262d9b36042c15444047aae226e#egg=dynamixel_sdk&subdirectory=../../../third_party/DynamixelSDK/python
easydict==1.9
easyhid==0.0.10
einops==0.8.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.3.0/en_core_web_sm-3.3.0-py3-none-any.whl
etils==1.12.2
evdev==1.6.0
example-interfaces==0.9.3
examples-rclpy-executors==0.15.3
examples-rclpy-minimal-action-client==0.15.3
examples-rclpy-minimal-action-server==0.15.3
examples-rclpy-minimal-client==0.15.3
examples-rclpy-minimal-publisher==0.15.3
examples-rclpy-minimal-service==0.15.3
examples-rclpy-minimal-subscriber==0.15.3
exceptiongroup==1.2.2
executing==2.2.0
Faker==13.15.0
Farama-Notifications==0.0.4
fastrlock==0.8
filelock==3.8.0
Flask==2.2.2
flatbuffers==1.12
fonttools==4.33.3
fsspec==2024.6.1
gast==0.4.0
-e git+ssh://git@github.com/wuphilipp/gello_software.git@daae81f4a78ec0f7534937413345d96d3e1bc7fc#egg=gello
geometry-msgs==4.2.4
gitdb==4.0.9
GitPython==3.1.27
glfw==2.8.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
GPUtil==1.4.0
grpcio==1.46.3
gym-bandits @ git+https://github.com/JKCooper2/gym-bandits@417ed323ca2f7298a3abdad34b781fa9f13148f1
gym-gridworlds==0.0.2
h5py==3.12.1
huggingface-hub==0.9.1
idna==3.10
image-geometry==3.2.1
importlib-metadata==4.13.0
importlib-resources==5.7.1
iniconfig==1.1.1
interactive-markers==2.3.2
ipython==8.32.0
isort==5.10.1
itsdangerous==2.1.2
jax-jumpy==1.0.0
jedi==0.19.2
Jinja2==3.1.2
joblib==1.1.0
keras==2.9.0
Keras-Preprocessing==1.1.2
keyboard==0.13.5
kiwisolver==1.4.2
labmaze==1.0.6
langcodes==3.3.0
laser-geometry==2.4.0
launch==1.0.7
launch-ros==0.19.8
launch-testing==1.0.7
launch-testing-ros==0.19.8
launch-xml==1.0.7
launch-yaml==1.0.7
lazy-object-proxy==1.7.1
libclang==14.0.1
lifecycle-msgs==1.2.1
llvmlite==0.44.0
logging-demo==0.20.5
lxml==5.3.1
map-msgs==2.1.0
Markdown==3.3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.1
math3d==4.0.0
matplotlib==3.5.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
message-filters==4.3.5
minimalmodbus==2.1.1
mpmath==1.3.0
mujoco==3.3.0
murmurhash==1.0.7
# Editable install with no version control (my-robot-controller==0.0.0)
-e /home/nums/ros2_ws/build/my_robot_controller
mypy==0.971
mypy-extensions==0.4.3
nav-msgs==4.2.4
networkx==3.3
nltk==3.7
numba==0.61.0
numpy==1.26.0
numpy-quaternion==2024.0.7
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.10.0.84
opt-einsum==3.3.0
osrf-pycommon==2.1.4
packaging==21.3
pandas==1.4.2
parso==0.8.4
pathspec==0.10.1
pathtools==0.1.2
pathy==0.6.1
pcl-msgs==1.0.0
pendulum-msgs==0.20.5
pexpect==4.9.0
pillow==11.1.0
platformdirs==2.5.2
pluggy==1.0.0
preshed==3.0.6
promise==2.3
prompt_toolkit==3.0.50
protobuf==6.30.1
ptyprocess==0.7.0
pure-python-adb==0.3.0.dev0
pure_eval==0.2.3
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybullet==3.2.5
pycocoevalcap==1.2
pycocotools==2.0.5
pycodestyle==2.9.1
pycparser==2.22
pydantic==1.8.2
pyFirmata==1.1.0
pygame==2.6.1
pyglet==1.5.26
Pygments==2.19.1
pylint==2.15.3
pymodbus==3.6.9
pynput==1.7.7
PyOpenGL==3.1.9
pyparsing==3.2.1
PyQt6==6.8.1
PyQt6-Qt6==6.8.2
PyQt6_sip==13.10.0
pyquaternion==0.9.9
pyrealsense2==2.55.1.6486
-e git+https://github.com/markusgrotz/pyRobotiqGripper.git@3da6dfbddd366a73c0006fe5c14731f7cf7f9bea#egg=pyRobotiqGripper
pyserial==3.5
pyspacemouse==1.1.4
pytest==7.1.3
pytest-cov==3.0.0
python-dateutil==2.9.0.post0
python-qt-binding==1.1.2
python-xlib==0.33
pyzmq==26.3.0
qt-dotgraph==2.2.3
qt-gui==2.2.3
qt-gui-cpp==2.2.3
qt-gui-py-common==2.2.3
quality-of-service-demo-py==0.20.5
Quaternion==3.5.2.post4
rcl-interfaces==1.2.1
rclpy==3.3.15
rcutils==5.1.6
regex==2022.6.2
requests==2.32.3
requests-oauthlib==1.3.1
resource-retriever==3.1.2
rich==13.9.4
rmw-dds-common==1.6.0
robotiq-modbus-controller @ git+https://github.com/tingelst/robotiq_modbus_controller.git@2825f9e53ac101bdf336c4f7daef38cc374c0ddc
ros2action==0.18.11
ros2bag==0.15.13
ros2cli==0.18.11
ros2component==0.18.11
ros2doctor==0.18.11
ros2interface==0.18.11
ros2launch==0.19.8
ros2lifecycle==0.18.11
ros2multicast==0.18.11
ros2node==0.18.11
ros2param==0.18.11
ros2pkg==0.18.11
ros2run==0.18.11
ros2service==0.18.11
ros2topic==0.18.11
rosbag2-interfaces==0.15.13
rosbag2-py==0.15.13
rosgraph-msgs==1.2.1
rosidl-adapter==3.1.6
rosidl-cli==3.1.6
rosidl-cmake==3.1.6
rosidl-generator-c==3.1.6
rosidl-generator-cpp==3.1.6
rosidl-generator-py==0.14.4
rosidl-parser==3.1.6
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.2
rosidl-typesupport-cpp==2.0.2
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.6
rosidl-typesupport-introspection-cpp==3.1.6
rpyutils==0.2.1
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-bag-plugins==1.1.5
rqt-console==2.0.3
rqt-graph==1.3.1
rqt-gui==1.1.7
rqt-gui-py==1.1.7
rqt-msg==1.2.0
rqt-plot==1.1.3
rqt-publisher==1.5.0
rqt-py-common==1.1.7
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
rsa==4.8
sb3-contrib==1.7.0
scikit-learn==1.1.1
scipy==1.15.2
sensor-msgs==4.2.4
sensor-msgs-py==4.2.4
sentencepiece==0.1.96
sentry-sdk==1.9.8
setproctitle==1.3.2
shape-msgs==4.2.4
shapely==2.0.2
shortuuid==1.0.9
shtab==1.7.1
six==1.17.0
smart-open==5.2.1
smmap==5.0.0
spacy==3.3.0
spacy-legacy==3.0.9
spacy-loggers==1.0.2
sros2==0.10.5
srsly==2.4.3
stack-data==0.6.3
statistics-msgs==1.2.1
std-msgs==4.2.4
std-srvs==4.2.4
stereo-msgs==4.2.4
sympy==1.13.1
teleop-twist-keyboard==2.4.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
tf2-geometry-msgs==0.25.12
tf2-kdl==0.25.12
tf2-msgs==0.25.12
tf2-py==0.25.12
tf2-ros-py==0.25.12
tf2-tools==0.25.12
thinc==8.0.17
threadpoolctl==3.1.0
tokenizers==0.12.1
tomli==2.0.1
tomlkit==0.11.4
topic-monitor==0.20.5
torch==2.6.0+cu124
torchvision==0.21.0+cu124
tqdm==4.64.0
traitlets==5.14.3
trajectory-msgs==4.2.4
transformers==4.22.1
triton==3.2.0
turtlesim==1.4.2
typeguard==4.4.2
typer==0.4.1
typing_extensions==4.12.2
tyro==0.9.17
unique-identifier-msgs==2.2.1
ur_rtde==1.6.0
urllib3==1.26.12
urx==0.11.0
visualization-msgs==4.2.4
wasabi==0.9.1
wcwidth==0.2.13
Werkzeug==2.2.2
wrapt==1.14.1
xarm==0.0.4
xArm-Python-SDK==1.14.8
youtube-dl==2021.12.17
zipp==3.21.0
zmq==0.0.0
